Made Fly work for Borderlands 3
Applied knowledge learned from Pwn Adventure Fly version, converted it to 64-bit asm for Borderlands 3, Found places to hook to get PitchYawPointer and PlayerPositionPointer, and as an improvement added a third hook to nullify gravity (as in this game it's pretty intense so it's kind of need to make flying comfortable). It does so by hooking the code that writes to player and other entity coordinates and filtering for only when you're flying and only when it's trying to write to the player coordinates then it only allows it to write the x and y coordinates excluding z.